+++ src/2-two/contraction-two/compute/isPossessive.js
@@ -7,10 +7,12 @@ const banList = {
 }
 
 const beforePossessive = {
-  in: true,//in sunday's
-  by: true,//by sunday's
-  for: true,//for sunday's
+  in: true, //in sunday's
+  by: true, //by sunday's
+  for: true, //for sunday's
 }
+let adjLike = new Set(['too', 'also', 'enough', 'about'])
+let nounLike = new Set(['is', 'are', 'did', 'were', 'could', 'should', 'must', 'had', 'have'])
 
 const isPossessive = (terms, i) => {
   let term = terms[i]
@@ -42,17 +44,59 @@ const isPossessive = (terms, i) => {
     }
     return false
   }
+  // the sun's setting vs the artist's painting
+  // gerund = is,  noun = possessive
+  // (we are doing some dupe-work of the switch classifier here)
+  if (nextTerm.switch == 'Noun|Gerund') {
+    let next2 = terms[i + 2]
+    // the artist's painting.
+    if (!next2) {
+      if (term.tags.has('Actor') || term.tags.has('ProperNoun')) {
+        return true
+      }
+      return false
+    }
+    // the artist's painting is..
+    if (next2.tags.has('Copula')) {
+      return true
+    }
+    // the cat's sleeping on ..
+    if (next2.normal === 'on' || next2.normal === 'in') {
+      return false
+    }
+    return false
+  }
   //a gerund suggests 'is walking'
   if (nextTerm.tags.has('Verb')) {
     //fix 'jamie's bite'
     if (nextTerm.tags.has('Infinitive')) {
       return true
     }
+    //'jamaica's growing'
+    if (nextTerm.tags.has('Gerund')) {
+      return false
+    }
     //fix 'spencer's runs'
     if (nextTerm.tags.has('PresentTense')) {
       return true
     }
     return false
+  }
+
+  // john's nuts
+  if (nextTerm.switch === 'Adj|Noun') {
+    let twoTerm = terms[i + 2]
+    if (!twoTerm) {
+      return false //adj
+    }
+    // john's nuts were..
+    if (nounLike.has(twoTerm.normal)) {
+      return true
+    }
+    // john's nuts about..
+    if (adjLike.has(twoTerm.normal)) {
+      return false //adj
+    }
   }
   //spencer's house
   if (nextTerm.tags.has('Noun')) {
@@ -65,25 +109,46 @@ const isPossessive = (terms, i) => {
     if (nextTerm.tags.has('Possessive')) {
       return false
     }
-    // the captain's John 
+    // the captain's John
     if (nextTerm.tags.has('ProperNoun') && !term.tags.has('ProperNoun')) {
       return false
     }
     return true
   }
+
   // by sunday's final
   if (terms[i - 1] && beforePossessive[terms[i - 1].normal] === true) {
     return true
   }
-  //rocket's red glare
-  let twoTerm = terms[i + 2]
-  if (twoTerm && twoTerm.tags.has('Noun') && !twoTerm.tags.has('Pronoun')) {
-    return true
-  }
-  //othwerwise, an adjective suggests 'is good'
-  if (nextTerm.tags.has('Adjective') || nextTerm.tags.has('Adverb') || nextTerm.tags.has('Verb')) {
+
+  // spencer's tired
+  if (nextTerm.tags.has('Adjective')) {
+    let twoTerm = terms[i + 2]
+    //the rocket's red
+    if (!twoTerm) {
+      return false
+    }
+    // rocket's red nozzle
+    if (twoTerm.tags.has('Noun') && !twoTerm.tags.has('Pronoun')) {
+      //project's behind schedule
+      let str = nextTerm.normal
+      if (str === 'above' || str === 'below' || str === 'behind') {
+        return false
+      }
+      return true
+    }
+    // rocket's red glare
+    if (twoTerm.switch === 'Noun|Verb') {
+      return true
+    }
+    //othwerwise, an adjective suggests 'is good'
     return false
   }
+  // baby's first steps
+  if (nextTerm.tags.has('Value')) {
+    return true
+  }
+  // otherwise not possessive
   return false
 }
 export default isPossessive