+++ src/2-two/postTagger/model/adjective/adj-gerund.js
@@ -13,13 +13,27 @@ export default [
   // a bit amusing
   { match: 'a (little|bit|wee) bit? [#Gerund]', group: 0, tag: 'Adjective', reason: 'a-bit-gerund' },
   // looking annoying
-  { match: '#Gerund [#Gerund]', group: 0, tag: 'Adjective', notIf: '(impersonating|practicing|considering|assuming)', reason: 'looking-annoying' },
+  {
+    match: '#Gerund [#Gerund]',
+    group: 0,
+    tag: 'Adjective',
+    notIf: '(impersonating|practicing|considering|assuming)',
+    reason: 'looking-annoying',
+  },
   // looked amazing
-  { match: '(looked|look|looks) #Adverb? [%Adj|Gerund%]', group: 0, tag: 'Adjective', notIf: '(impersonating|practicing|considering|assuming)', reason: 'looked-amazing' },
+  {
+    match: '(looked|look|looks) #Adverb? [%Adj|Gerund%]',
+    group: 0,
+    tag: 'Adjective',
+    notIf: '(impersonating|practicing|considering|assuming)',
+    reason: 'looked-amazing',
+  },
   // were really amazing
   // { match: '(looked|look|looks) #Adverb [%Adj|Gerund%]', group: 0, tag: 'Adjective', notIf: '(impersonating|practicing|considering|assuming)', reason: 'looked-amazing' },
   // developing a
   { match: '[%Adj|Gerund%] #Determiner', group: 0, tag: 'Gerund', reason: 'developing-a' },
   // world's leading manufacturer
   { match: '#Possessive [%Adj|Gerund%] #Noun', group: 0, tag: 'Adjective', reason: 'leading-manufacturer' },
+  // meaning alluring
+  { match: '%Noun|Gerund% %Adj|Gerund%', tag: 'Gerund #Adjective', reason: 'meaning-alluring' },
 ]